add two different SystemInfoService
